Bot/requirements.txt
gradio
mistralai
python-dotenv 